agenteai/requirements.txt
openai>=1.10.0
python-dotenv==0.19.0 